+++ src/dist_s1/data_models/runconfig_model.py
@@ -134,7 +134,7 @@ class RunConfigData(BaseModel):
     )
     tqdm_enabled: bool = Field(default=True)
     batch_size_for_despeckling: int = Field(
-        default=100,
+        default=25,
         ge=1,
     )
     n_workers_for_norm_param_estimation: int = Field(