+++ src/dist_s1/workflows.py
@@ -1,4 +1,3 @@
-import multiprocessing as mp
 from datetime import datetime
 from functools import partial
 from pathlib import Path