+++ examples/e2e.sh
@@ -3,4 +3,7 @@ dist-s1 run \
     --post_date '2025-01-21' \
     --track_number 71 \
     --dst_dir '../notebooks/los-angeles' \
-    --memory_strategy 'high' 
+    --memory_strategy 'low' \
+    --device 'cpu' \
+    --n_workers_for_norm_param_estimation 5 \
+    --batch_size_for_despeckling 100