+++ src/dist_s1/processing.py
@@ -53,8 +53,14 @@ def compute_normal_params_per_burst_and_serialize(
     out_path_sigma_copol: Path,
     out_path_sigma_crosspol: Path,
     memory_strategy: str = 'high',
+    device: str = 'best',
 ) -> Path:
-    model = load_transformer_model()
+    if device not in ('cpu', 'cuda', 'mps', 'best'):
+        raise ValueError(f'Invalid device: {device}')
+    # For distmetrics, None is how we choose the "best" available device
+    if device == 'best':
+        device = None
+    model = load_transformer_model(device=device)
 
     copol_data = [open_one_ds(path) for path in pre_copol_paths_dskpl_paths]
     crosspol_data = [open_one_ds(path) for path in pre_crosspol_paths_dskpl_paths]
@@ -69,7 +75,7 @@ def compute_normal_params_per_burst_and_serialize(
         check_profiles_match(p_ref, p_crosspol)
 
     logits_mu, logits_sigma = estimate_normal_params_of_logits(
-        model, arrs_copol, arrs_crosspol, memory_strategy=memory_strategy
+        model, arrs_copol, arrs_crosspol, memory_strategy=memory_strategy, device=device
     )
     logits_mu_copol, logits_mu_crosspol = logits_mu[0, ...], logits_mu[1, ...]
     logits_sigma_copol, logits_sigma_crosspol = logits_sigma[0, ...], logits_sigma[1, ...]
@@ -78,10 +84,6 @@ def compute_normal_params_per_burst_and_serialize(
     serialize_one_2d_ds(logits_mu_crosspol, p_ref, out_path_mu_crosspol)
     serialize_one_2d_ds(logits_sigma_copol, p_ref, out_path_sigma_copol)
     serialize_one_2d_ds(logits_sigma_crosspol, p_ref, out_path_sigma_crosspol)
-
-
-def compute_disturbance(metric_paths: list[Path], out_dir: Path) -> None:
-    pass
 
 
 def compute_logit_mdist(arr_logit: np.ndarray, mean_logit: np.ndarray, sigma_logit: np.ndarray) -> np.ndarray: