+++ src/dist_s1/workflows.py
@@ -1,7 +1,9 @@
 from datetime import datetime
+from functools import partial
 from pathlib import Path
 
 import pandas as pd
+import torch.multiprocessing as tmp
 from tqdm.auto import tqdm
 
 from dist_s1.aws import upload_product_to_s3
@@ -17,6 +19,10 @@ from dist_s1.processing import (
     merge_burst_disturbances_and_serialize,
     merge_burst_metrics_and_serialize,
 )
+
+
+# Use spawn for multiprocessing
+tmp.set_start_method('spawn', force=True)
 
 
 def curate_input_burst_rtc_input_for_dist(
@@ -54,6 +60,51 @@ def curate_input_burst_rtc_s1_paths_for_normal_param_est(
     copol_paths_pre = copol_paths[start:stop]
     crosspol_paths_pre = crosspol_paths[start:stop]
     return copol_paths_pre, crosspol_paths_pre
+
+
+def curate_paths_for_normal_param_est_via_burst_id_and_lookback(
+    df_inputs: pd.DataFrame, df_burst_distmetrics: pd.DataFrame, burst_id: str, lookback: int
+) -> tuple[list[Path], list[Path]]:
+    """Curate the paths to the correct length for normal param estimation."""
+    indices_input = (df_inputs.jpl_burst_id == burst_id) & (df_inputs.input_category == 'pre')
+    df_burst_input_data = df_inputs[indices_input].reset_index(drop=True)
+    indices_input = (df_inputs.jpl_burst_id == burst_id) & (df_inputs.input_category == 'pre')
+    df_burst_input_data = df_inputs[indices_input].reset_index(drop=True)
+    df_metric = df_burst_distmetrics[df_burst_distmetrics.jpl_burst_id == burst_id].reset_index(drop=True)
+
+    copol_paths = df_burst_input_data.loc_path_copol_dspkl.tolist()
+    crosspol_paths = df_burst_input_data.loc_path_crosspol_dspkl.tolist()
+
+    # curate the paths to the correct length
+    copol_paths_pre, crosspol_paths_pre = curate_input_burst_rtc_s1_paths_for_normal_param_est(
+        copol_paths, crosspol_paths, lookback
+    )
+
+    output_mu_copol_l = df_metric[f'loc_path_normal_mean_delta{lookback}_copol'].tolist()
+    output_mu_crosspol_l = df_metric[f'loc_path_normal_mean_delta{lookback}_crosspol'].tolist()
+    output_sigma_copol_l = df_metric[f'loc_path_normal_std_delta{lookback}_copol'].tolist()
+    output_sigma_crosspol_l = df_metric[f'loc_path_normal_std_delta{lookback}_crosspol'].tolist()
+
+    assert (
+        len(output_mu_copol_l)
+        == len(output_mu_crosspol_l)
+        == len(output_sigma_copol_l)
+        == len(output_sigma_crosspol_l)
+        == 1
+    )
+
+    output_mu_copol_path = output_mu_copol_l[0]
+    output_mu_crosspol_path = output_mu_crosspol_l[0]
+    output_sigma_copol_path = output_sigma_copol_l[0]
+    output_sigma_crosspol_path = output_sigma_crosspol_l[0]
+    return dict(
+        copol_paths_pre=copol_paths_pre,
+        crosspol_paths_pre=crosspol_paths_pre,
+        output_mu_copol_path=output_mu_copol_path,
+        output_mu_crosspol_path=output_mu_crosspol_path,
+        output_sigma_copol_path=output_sigma_copol_path,
+        output_sigma_crosspol_path=output_sigma_crosspol_path,
+    )
 
 
 def check_dates_of_normal_param_files(pre_paths: list[str], normal_out_path: Path) -> bool:
@@ -120,7 +171,25 @@ def run_despeckle_workflow(run_config: RunConfigData) -> None:
     rtc_paths = copol_paths + crosspol_paths
     dst_paths = dspkl_copol_paths + dspkl_crosspol_paths
 
-    despeckle_and_serialize_rtc_s1(rtc_paths, dst_paths, n_workers=run_config.n_workers_for_despeckling)
+    despeckle_and_serialize_rtc_s1(
+        rtc_paths,
+        dst_paths,
+        n_workers=run_config.n_workers_for_despeckling,
+        batch_size=run_config.batch_size_for_despeckling,
+    )
+
+
+def _process_normal_params(path_data: dict, memory_strategy: str, device: str) -> None:
+    return compute_normal_params_per_burst_and_serialize(
+        path_data['copol_paths_pre'],
+        path_data['crosspol_paths_pre'],
+        path_data['output_mu_copol_path'],
+        path_data['output_mu_crosspol_path'],
+        path_data['output_sigma_copol_path'],
+        path_data['output_sigma_crosspol_path'],
+        memory_strategy=memory_strategy,
+        device=device,
+    )
 
 
 def run_normal_param_estimation_workflow(run_config: RunConfigData) -> None:
@@ -134,62 +203,53 @@ def run_normal_param_estimation_workflow(run_config: RunConfigData) -> None:
     df_burst_distmetrics = run_config.df_burst_distmetrics
 
     tqdm_disable = not run_config.tqdm_enabled
-    for burst_id in tqdm(
-        df_inputs.jpl_burst_id.unique(), disable=tqdm_disable, desc='Param Est. by Burst', dynamic_ncols=True
-    ):
-        for lookback in tqdm(
-            range(run_config.n_lookbacks),
+    burst_id_lookback_pairs = [
+        (burst_id, lookback)
+        for burst_id in df_inputs.jpl_burst_id.unique()
+        for lookback in range(run_config.n_lookbacks)
+    ]
+    norm_param_paths = [
+        curate_paths_for_normal_param_est_via_burst_id_and_lookback(df_inputs, df_burst_distmetrics, burst_id, lookback)
+        for burst_id, lookback in burst_id_lookback_pairs
+    ]
+
+    if run_config.n_workers_for_norm_param_estimation == 1:
+        for path_data in tqdm(
+            norm_param_paths,
             disable=tqdm_disable,
-            desc=f'Lookbacks for burst {burst_id}',
+            desc='Normal param estimation for burst/lookback pairs',
             dynamic_ncols=True,
             leave=False,
         ):
-            indices_input = (df_inputs.jpl_burst_id == burst_id) & (df_inputs.input_category == 'pre')
-            df_burst_input_data = df_inputs[indices_input].reset_index(drop=True)
-            df_metric = df_burst_distmetrics[df_burst_distmetrics.jpl_burst_id == burst_id].reset_index(drop=True)
-
-            copol_paths = df_burst_input_data.loc_path_copol_dspkl.tolist()
-            crosspol_paths = df_burst_input_data.loc_path_crosspol_dspkl.tolist()
-
-            # curate the paths to the correct length
-            copol_paths_pre, crosspol_paths_pre = curate_input_burst_rtc_s1_paths_for_normal_param_est(
-                copol_paths, crosspol_paths, lookback
-            )
-
-            output_mu_copol_l = df_metric[f'loc_path_normal_mean_delta{lookback}_copol'].tolist()
-            output_mu_crosspol_l = df_metric[f'loc_path_normal_mean_delta{lookback}_crosspol'].tolist()
-            output_sigma_copol_l = df_metric[f'loc_path_normal_std_delta{lookback}_copol'].tolist()
-            output_sigma_crosspol_l = df_metric[f'loc_path_normal_std_delta{lookback}_crosspol'].tolist()
-
-            assert (
-                len(output_mu_copol_l)
-                == len(output_mu_crosspol_l)
-                == len(output_sigma_copol_l)
-                == len(output_sigma_crosspol_l)
-                == 1
-            )
-
-            output_mu_copol_path = output_mu_copol_l[0]
-            output_mu_crosspol_path = output_mu_crosspol_l[0]
-            output_sigma_copol_path = output_sigma_copol_l[0]
-            output_sigma_crosspol_path = output_sigma_crosspol_l[0]
-
-            checks = [
-                check_dates_of_normal_param_files(copol_paths_pre, output_mu_copol_path),
-                check_dates_of_normal_param_files(crosspol_paths_pre, output_mu_crosspol_path),
-                check_dates_of_normal_param_files(copol_paths_pre, output_sigma_copol_path),
-                check_dates_of_normal_param_files(crosspol_paths_pre, output_sigma_crosspol_path),
-            ]
-            assert all(checks)
-
             compute_normal_params_per_burst_and_serialize(
-                copol_paths_pre,
-                crosspol_paths_pre,
-                output_mu_copol_path,
-                output_mu_crosspol_path,
-                output_sigma_copol_path,
-                output_sigma_crosspol_path,
+                path_data['copol_paths_pre'],
+                path_data['crosspol_paths_pre'],
+                path_data['output_mu_copol_path'],
+                path_data['output_mu_crosspol_path'],
+                path_data['output_sigma_copol_path'],
+                path_data['output_sigma_crosspol_path'],
                 memory_strategy=run_config.memory_strategy,
+                device=run_config.device,
+            )
+    else:
+        if run_config.device in ('cuda', 'mps'):
+            raise NotImplementedError('Multi-GPU processing is not supported yet')
+
+        # Create a partial function with the memory strategy and device
+        worker_fn = partial(
+            _process_normal_params, memory_strategy=run_config.memory_strategy, device=run_config.device
+        )
+
+        # Start a pool of workers
+        with tmp.Pool(processes=run_config.n_workers_for_norm_param_estimation) as pool:
+            # Map the work to the pool and show progress
+            list(
+                tqdm(
+                    pool.imap(worker_fn, norm_param_paths),
+                    total=len(norm_param_paths),
+                    desc='Normal param estimation for burst/lookback pairs',
+                    dynamic_ncols=True,
+                )
             )
 
 
@@ -317,6 +377,9 @@ def run_dist_s1_sas_prep_workflow(
     bucket: str | None = None,
     bucket_prefix: str = '',
     n_workers_for_despeckling: int = 5,
+    device: str = 'best',
+    batch_size_for_despeckling: int = 25,
+    n_workers_for_norm_param_estimation: int = 1,
 ) -> RunConfigData:
     run_config = run_dist_s1_localization_workflow(
         mgrs_tile_id,
@@ -339,6 +402,9 @@ def run_dist_s1_sas_prep_workflow(
     run_config.bucket = bucket
     run_config.bucket_prefix = bucket_prefix
     run_config.n_workers_for_despeckling = n_workers_for_despeckling
+    run_config.batch_size_for_despeckling = batch_size_for_despeckling
+    run_config.n_workers_for_norm_param_estimation = n_workers_for_norm_param_estimation
+    run_config.device = device
     return run_config
 
 
@@ -370,6 +436,9 @@ def run_dist_s1_workflow(
     bucket: str | None = None,
     bucket_prefix: str = '',
     n_workers_for_despeckling: int = 5,
+    batch_size_for_despeckling: int = 25,
+    n_workers_for_norm_param_estimation: int = 1,
+    device: str = 'best',
 ) -> Path:
     run_config = run_dist_s1_sas_prep_workflow(
         mgrs_tile_id,
@@ -389,6 +458,9 @@ def run_dist_s1_workflow(
         bucket=bucket,
         bucket_prefix=bucket_prefix,
         n_workers_for_despeckling=n_workers_for_despeckling,
+        batch_size_for_despeckling=batch_size_for_despeckling,
+        n_workers_for_norm_param_estimation=n_workers_for_norm_param_estimation,
+        device=device,
     )
     _ = run_dist_s1_sas_workflow(run_config)